PB - update bo

+++ src/main/java/bo/Transport.java
@@ -6,6 +6,6 @@ import enums.Color;
 public abstract class Transport {
     protected Brand brand;
     protected Color color;
-    protected int fuelConsumption;
+    protected double fuelConsumption;
     protected double costPrice;
 }
